+++ src/constants.ts
@@ -1,4 +1,4 @@
-export const VERSION = "4.0.0";
+export const VERSION = "4.0.1";
 
 export const Status = {
   QUEUED: "queued",